pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opc"
version = "0.1.0"
description = "One Person Company - 单人软件公司 AI 系统"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.39",
    "python-dotenv",
    "rich",
    "chromadb>=0.4",
    "rank-bm25>=0.2",
    "jieba>=0.42",
    "tomli>=2; python_version < '3.11'",
]

[project.scripts]
opc = "opc.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
